refactor: 스토리 에디터 컴포넌트 구조 개선
