+++ src/app/restaurant/[id]/review/graphic/story-editor/_components/EditorHeader.tsx
